+++ internal/table/client.go
@@ -24,19 +24,19 @@ import (
 )
 
 var (
-	// errAlreadyClosed returned by a client instance to indicate
-	// that client is closed early and not able to complete requested operation.
-	errAlreadyClosed = xerrors.Wrap(fmt.Errorf("table client closed early"))
+	// errAlreadyClosed returned by a Client instance to indicate
+	// that Client is closed early and not able to complete requested operation.
+	errAlreadyClosed = xerrors.Wrap(fmt.Errorf("table Client closed early"))
 
-	// errSessionPoolOverflow returned by a client instance to indicate
-	// that the client is full and requested operation is not able to complete.
+	// errSessionPoolOverflow returned by a Client instance to indicate
+	// that the Client is full and requested operation is not able to complete.
 	errSessionPoolOverflow = xerrors.Wrap(fmt.Errorf("session pool overflow"))
 
-	// errSessionShutdown returned by a client instance to indicate that
+	// errSessionShutdown returned by a Client instance to indicate that
 	// requested session is under shutdown.
 	errSessionShutdown = xerrors.Wrap(fmt.Errorf("session under shutdown"))
 
-	// errNoProgress returned by a client instance to indicate that
+	// errNoProgress returned by a Client instance to indicate that
 	// operation could not be completed.
 	errNoProgress = xerrors.Wrap(fmt.Errorf("no progress"))
 )
@@ -44,15 +44,7 @@ var (
 // SessionBuilder is the interface that holds logic of creating sessions.
 type SessionBuilder func(context.Context) (Session, error)
 
-type Client interface {
-	table.Client
-
-	Get(ctx context.Context) (s Session, err error)
-	Put(ctx context.Context, s Session) (err error)
-	CloseSession(ctx context.Context, s Session) (err error)
-}
-
-func New(cc grpc.ClientConnInterface, config config.Config) Client {
+func New(cc grpc.ClientConnInterface, config config.Config) *Client {
 	return newClient(cc, nil, config)
 }
 
@@ -60,7 +52,7 @@ func newClient(
 	cc grpc.ClientConnInterface,
 	builder SessionBuilder,
 	config config.Config,
-) *client {
+) *Client {
 	var (
 		ctx    = context.Background()
 		onDone = trace.TableOnInit(config.Trace(), &ctx)
@@ -70,7 +62,7 @@ func newClient(
 			return newSession(ctx, cc, config)
 		}
 	}
-	c := &client{
+	c := &Client{
 		config: config,
 		cc:     cc,
 		build:  builder,
@@ -94,9 +86,9 @@ func newClient(
 	return c
 }
 
-// client is a set of session instances that may be reused.
-// A client is safe for use by multiple goroutines simultaneously.
-type client struct {
+// Client is a set of session instances that may be reused.
+// A Client is safe for use by multiple goroutines simultaneously.
+type Client struct {
 	// build holds an object capable for creating sessions.
 	// It must not be nil.
 	build             SessionBuilder
@@ -104,7 +96,7 @@ type client struct {
 	config            config.Config
 	index             map[Session]sessionInfo
 	createInProgress  int           // KIKIMR-9163: in-create-process counter
-	limit             int           // Upper bound for client size.
+	limit             int           // Upper bound for Client size.
 	idle              *list.List    // list<table.session>
 	waitq             *list.List    // list<*chan table.session>
 	keeperWake        chan struct{} // Set by keeper.
@@ -119,7 +111,7 @@ type client struct {
 	closed            bool
 }
 
-func (c *client) CreateSession(ctx context.Context, opts ...table.Option) (table.ClosableSession, error) {
+func (c *Client) CreateSession(ctx context.Context, opts ...table.Option) (table.ClosableSession, error) {
 	var (
 		s       Session
 		err     error
@@ -150,7 +142,7 @@ func (c *client) CreateSession(ctx context.Context, opts ...table.Option) (table
 	return s, xerrors.WithStackTrace(err)
 }
 
-func (c *client) isClosed() bool {
+func (c *Client) isClosed() bool {
 	c.mu.Lock()
 	defer c.mu.Unlock()
 	return c.closed
@@ -189,8 +181,8 @@ type createSessionResult struct {
 }
 
 // p.mu must NOT be held.
-func (c *client) createSession(ctx context.Context) (s Session, err error) {
-	// pre-check the client size
+func (c *Client) createSession(ctx context.Context) (s Session, err error) {
+	// pre-check the Client size
 	c.mu.Lock()
 	enoughSpace := c.createInProgress+len(c.index) < c.limit
 	if enoughSpace {
@@ -249,7 +241,7 @@ func (c *client) createSession(ctx context.Context) (s Session, err error) {
 				c.notify(nil)
 
 				if info.idle != nil {
-					panic("session closed while still in idle client")
+					panic("session closed while still in idle Client")
 				}
 			})
 		}
@@ -273,7 +265,7 @@ func (c *client) createSession(ctx context.Context) (s Session, err error) {
 	select {
 	case r := <-resCh:
 		if r.s == nil && r.err == nil {
-			panic("ydb: abnormal result of client.createSession()")
+			panic("ydb: abnormal result of Client.createSession()")
 		}
 		return r.s, xerrors.WithStackTrace(r.err)
 	case <-ctx.Done():
@@ -299,7 +291,7 @@ func withTrace(t trace.Table) getOption {
 	}
 }
 
-func (c *client) get(ctx context.Context, opts ...getOption) (s Session, err error) {
+func (c *Client) get(ctx context.Context, opts ...getOption) (s Session, err error) {
 	var (
 		i = 0
 		o = getOptions{t: c.config.Trace()}
@@ -338,7 +330,7 @@ func (c *client) get(ctx context.Context, opts ...getOption) (s Session, err err
 			return s, xerrors.WithStackTrace(err)
 		}
 
-		// Third, we try to wait for a touched session - client is full.
+		// Third, we try to wait for a touched session - Client is full.
 		//
 		// This should be done only if number of currently waiting goroutines
 		// are less than maximum amount of touched session. That is, we want to
@@ -360,13 +352,13 @@ func (c *client) get(ctx context.Context, opts ...getOption) (s Session, err err
 	return s, nil
 }
 
-// Get returns first idle session from the client and removes it from
-// there. If no items stored in client it creates new one returns it.
-func (c *client) Get(ctx context.Context) (s Session, err error) {
+// Get returns first idle session from the Client and removes it from
+// there. If no items stored in Client it creates new one returns it.
+func (c *Client) Get(ctx context.Context) (s Session, err error) {
 	return c.get(ctx)
 }
 
-func (c *client) waitFromCh(ctx context.Context, t trace.Table) (s Session, err error) {
+func (c *Client) waitFromCh(ctx context.Context, t trace.Table) (s Session, err error) {
 	var (
 		ch *chan Session
 		el *list.Element // Element in the wait queue.
@@ -393,7 +385,7 @@ func (c *client) waitFromCh(ctx context.Context, t trace.Table) (s Session, err 
 		// The same way will work when some session become deleted - the
 		// nil value will be sent into the channel.
 		if ok {
-			// Put only filled and not closed channel back to the client.
+			// Put only filled and not closed channel back to the Client.
 			// That is, we need to avoid races on filling reused channel
 			// for the next waiter – session could be lost for a long time.
 			c.putWaitCh(ch)
@@ -414,16 +406,16 @@ func (c *client) waitFromCh(ctx context.Context, t trace.Table) (s Session, err 
 	}
 }
 
-// Put returns session to the client for further reuse.
-// If client is already closed Put() calls s.Close(ctx) and returns
+// Put returns session to the Client for further reuse.
+// If Client is already closed Put() calls s.Close(ctx) and returns
 // errAlreadyClosed.
-// If client is overflow calls s.Close(ctx) and returns
+// If Client is overflow calls s.Close(ctx) and returns
 // errSessionPoolOverflow.
 //
 // Note that Put() must be called only once after being created or received by
 // Get() or Take() calls. In other way it will produce unexpected behavior or
 // panic.
-func (c *client) Put(ctx context.Context, s Session) (err error) {
+func (c *Client) Put(ctx context.Context, s Session) (err error) {
 	onDone := trace.TableOnPoolPut(c.config.Trace(), &ctx, s)
 	defer func() {
 		onDone(err)
@@ -463,11 +455,11 @@ func (c *client) Put(ctx context.Context, s Session) (err error) {
 	return xerrors.WithStackTrace(err)
 }
 
-// Close deletes all stored sessions inside client.
+// Close deletes all stored sessions inside Client.
 // It also stops all underlying timers and goroutines.
 // It returns first error occurred during stale sessions' deletion.
 // Note that even on error it calls Close() on each session.
-func (c *client) Close(ctx context.Context) (err error) {
+func (c *Client) Close(ctx context.Context) (err error) {
 	onDone := trace.TableOnClose(c.config.Trace(), &ctx)
 	defer func() {
 		onDone(err)
@@ -516,7 +508,7 @@ func (c *client) Close(ctx context.Context) (err error) {
 	c.wgClosed.Wait()
 
 	if len(issues) > 0 {
-		return xerrors.WithStackTrace(xerrors.NewWithIssues("table client closed with issues", issues...))
+		return xerrors.WithStackTrace(xerrors.NewWithIssues("table Client closed with issues", issues...))
 	}
 
 	return nil
@@ -539,7 +531,7 @@ func retryOptions(trace trace.Table, opts ...table.Option) table.Options {
 // - deadline was canceled or deadlined
 // - retry operation returned nil as error
 // Warning: if deadline without deadline or cancellation func Retry will be worked infinite
-func (c *client) Do(ctx context.Context, op table.Operation, opts ...table.Option) (err error) {
+func (c *Client) Do(ctx context.Context, op table.Operation, opts ...table.Option) (err error) {
 	if c.isClosed() {
 		return xerrors.WithStackTrace(errAlreadyClosed)
 	}
@@ -553,7 +545,7 @@ func (c *client) Do(ctx context.Context, op table.Operation, opts ...table.Optio
 	)
 }
 
-func (c *client) DoTx(ctx context.Context, op table.TxOperation, opts ...table.Option) (err error) {
+func (c *Client) DoTx(ctx context.Context, op table.TxOperation, opts ...table.Option) (err error) {
 	if c.isClosed() {
 		return xerrors.WithStackTrace(errAlreadyClosed)
 	}
@@ -566,7 +558,7 @@ func (c *client) DoTx(ctx context.Context, op table.TxOperation, opts ...table.O
 	)
 }
 
-func (c *client) keeper(ctx context.Context) {
+func (c *Client) keeper(ctx context.Context) {
 	defer close(c.keeperDone)
 	var (
 		toTouch    []Session // Cached for reuse.
@@ -726,7 +718,7 @@ func (c *client) keeper(ctx context.Context) {
 // Note that returning a pointer reduces allocations on sync.Pool usage –
 // sync.Client.Get() returns empty interface, which leads to allocation for
 // non-pointer values.
-func (c *client) getWaitCh() *chan Session {
+func (c *Client) getWaitCh() *chan Session {
 	if c.testHookGetWaitCh != nil {
 		c.testHookGetWaitCh()
 	}
@@ -742,12 +734,12 @@ func (c *client) getWaitCh() *chan Session {
 // use.
 // Note that ch MUST NOT be owned by any goroutine at the call moment and ch
 // MUST NOT contain any value.
-func (c *client) putWaitCh(ch *chan Session) {
+func (c *Client) putWaitCh(ch *chan Session) {
 	c.waitChPool.Put(ch)
 }
 
 // p.mu must be held.
-func (c *client) peekFirstIdle() (s Session, touched time.Time) {
+func (c *Client) peekFirstIdle() (s Session, touched time.Time) {
 	el := c.idle.Front()
 	if el == nil {
 		return
@@ -755,7 +747,7 @@ func (c *client) peekFirstIdle() (s Session, touched time.Time) {
 	s = el.Value.(Session)
 	info, has := c.index[s]
 	if !has || el != info.idle {
-		panicLocked(&c.mu, "inconsistent session client index")
+		panicLocked(&c.mu, "inconsistent session Client index")
 	}
 	return s, info.touched
 }
@@ -764,7 +756,7 @@ func (c *client) peekFirstIdle() (s Session, touched time.Time) {
 // to prevent session from dying in the keeper after it was returned
 // to be used only in outgoing functions that make session busy.
 // p.mu must be held.
-func (c *client) removeFirstIdle() Session {
+func (c *Client) removeFirstIdle() Session {
 	s, _ := c.peekFirstIdle()
 	if s != nil {
 		info := c.removeIdle(s)
@@ -778,7 +770,7 @@ func (c *client) removeFirstIdle() Session {
 // Unlike other info modifiers, this one doesn't care if it didn't find the session, it skips
 // the action. You can still check it later if needed, if the return code is -1
 // p.mu must be held.
-func (c *client) incrementKeepAlive(s Session) int {
+func (c *Client) incrementKeepAlive(s Session) int {
 	info, has := c.index[s]
 	if !has {
 		return -1
@@ -790,7 +782,7 @@ func (c *client) incrementKeepAlive(s Session) int {
 }
 
 // p.mu must be held.
-func (c *client) touchCond() <-chan struct{} {
+func (c *Client) touchCond() <-chan struct{} {
 	if c.touchingDone == nil {
 		c.touchingDone = make(chan struct{})
 	}
@@ -798,7 +790,7 @@ func (c *client) touchCond() <-chan struct{} {
 }
 
 // p.mu must be held.
-func (c *client) notify(s Session) (notified bool) {
+func (c *Client) notify(s Session) (notified bool) {
 	for el := c.waitq.Front(); el != nil; el = c.waitq.Front() {
 		// Some goroutine is waiting for a session.
 		//
@@ -829,7 +821,7 @@ func (c *client) notify(s Session) (notified bool) {
 // CloseSession provides the most effective way of session closing
 // instead of plain session.Close().
 // CloseSession must be fast. If necessary, can be async.
-func (c *client) CloseSession(ctx context.Context, s Session) error {
+func (c *Client) CloseSession(ctx context.Context, s Session) error {
 	return c.closeSession(
 		ctx,
 		s,
@@ -866,7 +858,7 @@ func withCloseSessionTrace() closeSessionOption {
 }
 
 // closeSession is an async func which close session, but without `trace.OnPoolSessionClose` tracing
-func (c *client) closeSession(ctx context.Context, s Session, opts ...closeSessionOption) error {
+func (c *Client) closeSession(ctx context.Context, s Session, opts ...closeSessionOption) error {
 	h := closeSessionOptionsHolder{}
 
 	for _, o := range opts {
@@ -907,7 +899,7 @@ func (c *client) closeSession(ctx context.Context, s Session, opts ...closeSessi
 	return nil
 }
 
-func (c *client) keepAliveSession(ctx context.Context, s Session) (err error) {
+func (c *Client) keepAliveSession(ctx context.Context, s Session) (err error) {
 	ctx, cancel := context.WithTimeout(ctx, c.config.KeepAliveTimeout())
 	defer cancel()
 	err = s.KeepAlive(ctx)
@@ -918,10 +910,10 @@ func (c *client) keepAliveSession(ctx context.Context, s Session) (err error) {
 }
 
 // p.mu must be held.
-func (c *client) removeIdle(s Session) sessionInfo {
+func (c *Client) removeIdle(s Session) sessionInfo {
 	info, has := c.index[s]
 	if !has || info.idle == nil {
-		panicLocked(&c.mu, "inconsistent session client index")
+		panicLocked(&c.mu, "inconsistent session Client index")
 	}
 
 	c.idle.Remove(info.idle)
@@ -931,12 +923,12 @@ func (c *client) removeIdle(s Session) sessionInfo {
 }
 
 // p.mu must be held.
-func (c *client) pushIdle(s Session, now time.Time) {
+func (c *Client) pushIdle(s Session, now time.Time) {
 	c.handlePushIdle(s, now, c.idle.PushBack(s))
 }
 
 // p.mu must be held.
-func (c *client) pushIdleInOrder(s Session, now time.Time) (el *list.Element) {
+func (c *Client) pushIdleInOrder(s Session, now time.Time) (el *list.Element) {
 	var prev *list.Element
 	for prev = c.idle.Back(); prev != nil; prev = prev.Prev() {
 		s := prev.Value.(Session)
@@ -955,7 +947,7 @@ func (c *client) pushIdleInOrder(s Session, now time.Time) (el *list.Element) {
 }
 
 // p.mu must be held.
-func (c *client) pushIdleInOrderAfter(s Session, now time.Time, mark *list.Element) *list.Element {
+func (c *Client) pushIdleInOrderAfter(s Session, now time.Time, mark *list.Element) *list.Element {
 	if mark != nil {
 		n := c.idle.Len()
 		el := c.idle.InsertAfter(s, mark)
@@ -969,13 +961,13 @@ func (c *client) pushIdleInOrderAfter(s Session, now time.Time, mark *list.Eleme
 }
 
 // p.mu must be held.
-func (c *client) handlePushIdle(s Session, now time.Time, el *list.Element) {
+func (c *Client) handlePushIdle(s Session, now time.Time, el *list.Element) {
 	info, has := c.index[s]
 	if !has {
-		panicLocked(&c.mu, "trying to store session created outside of the client")
+		panicLocked(&c.mu, "trying to store session created outside of the Client")
 	}
 	if info.idle != nil {
-		panicLocked(&c.mu, "inconsistent session client index")
+		panicLocked(&c.mu, "inconsistent session Client index")
 	}
 
 	info.touched = now
@@ -986,7 +978,7 @@ func (c *client) handlePushIdle(s Session, now time.Time, el *list.Element) {
 }
 
 // p.mu must be held.
-func (c *client) wakeUpKeeper() {
+func (c *Client) wakeUpKeeper() {
 	if wake := c.keeperWake; wake != nil {
 		c.keeperWake = nil
 		close(wake)